python-ai-server/requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
angle-emb==0.3.5
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
bitsandbytes==0.42.0
boltons==23.1.1
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.17.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
filelock==3.9.0
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
grpcio==1.62.1
grpcio-health-checking==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.25
langdetect==1.0.9
langsmith==0.1.5
lxml==5.1.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.3
mplcursors==0.5.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
onnx==1.15.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
peft==0.8.2
pikepdf==8.13.0
pillow==10.2.0
prettytable==3.10.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydyf==0.8.0
pyparsing==3.1.1
pypdf==4.0.2
PyPDF2==3.0.1
pypdfium2==4.27.0
pyphen==0.14.0
python-dateutil==2.8.2
python-doctr==0.7.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
reportlab==4.1.0
requests==2.31.0
safetensors==0.4.2
scipy==1.12.0
sentry-sdk==1.40.5
setproctitle==1.3.3
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.2
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tqdm==4.66.2
transformers==4.38.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
validators==0.22.0
wandb==0.16.3
wcwidth==0.2.13
weasyprint==61.0
weaviate-client==4.5.2
webencodings==0.5.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
zopfli==0.2.3
